driftpa/pyproject.toml
[project]
name = "driftpa"
version = "0.1.0"
description = "Personal assistant RL environment with schema drift, time pressure, and irreversible actions"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["driftpa*"]
